% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Randomly changing the order of obersverations}
\usage{
resample(data, ID, variables)
}
\arguments{
\item{data}{A data frame containing hierarchical data (e.g., time series data)}

\item{ID}{The grouping variable (e.g., to distinguish participants from each other)}

\item{variables}{A vector of variable names for which the order should be changed.}
}
\value{
The resampled data frame.
}
\description{
This function randomly resamples observations separately for variables and groups (e.g., individual participants)
so that meaningful information contained in the order of the observation gets lost, but each groups still has the same set of values.
}
\examples{
tib  <- dplyr::tibble(participant=rep(1:10, each=10),  # grouping variable
happiness = rnorm(100, mean = 3.5, sd=1.7),     # order should not be changed
stress = rnorm(100, mean = 2, sd= 1.3),         # order should be changed
anxiety = rnorm(100, mean = 1.7, sd= 2))        # order should be changed

new_tib <- resample(tib, participant, c("stress", "anxiety"))

}
